requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==24.1.0
    # via -r requirements.in
aiohappyeyeballs==2.4.0
    # via aiohttp
aiohttp==3.10.5
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==24.2.0
    # via aiohttp
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
coverage[toml]==7.6.1
    # via pytest-cov
dataclasses-json==0.6.7
    # via langchain-community
docstring-parser==0.16
    # via google-cloud-aiplatform
exceptiongroup==1.2.2
    # via
    #   anyio
    #   pytest
execnet==2.1.1
    # via pytest-cache
fastapi==0.114.0
    # via -r requirements.in
flake8==5.0.4
    # via pytest-flake8
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
google-api-core[grpc]==2.19.2
    # via
    #   google-api-python-client
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-discoveryengine
    #   google-cloud-resource-manager
    #   google-cloud-storage
    #   langchain-google-community
google-api-python-client==2.144.0
    # via langchain-google-community
google-auth==2.34.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-discoveryengine
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-cloud-aiplatform==1.65.0
    # via
    #   -r requirements.in
    #   langchain-google-vertexai
google-cloud-bigquery==3.25.0
    # via google-cloud-aiplatform
google-cloud-core==2.4.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-discoveryengine==0.12.2
    # via -r requirements.in
google-cloud-resource-manager==1.12.5
    # via google-cloud-aiplatform
google-cloud-storage==2.18.2
    # via
    #   google-cloud-aiplatform
    #   langchain-google-vertexai
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.65.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.13.1
    # via google-cloud-resource-manager
grpcio==1.66.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   langchain-google-community
grpcio-status==1.66.1
    # via google-api-core
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.6.1
    # via uvicorn
httpx==0.27.2
    # via
    #   langchain-google-vertexai
    #   langsmith
httpx-sse==0.4.0
    # via langchain-google-vertexai
idna==3.8
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via -r requirements.in
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.2.16
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.2.16
    # via langchain-google-community
langchain-core==0.2.38
    # via
    #   langchain
    #   langchain-community
    #   langchain-google-community
    #   langchain-google-vertexai
    #   langchain-text-splitters
langchain-google-community==1.0.8
    # via -r requirements.in
langchain-google-vertexai==1.0.10
    # via -r requirements.in
langchain-text-splitters==0.2.4
    # via langchain
langsmith==0.1.116
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markupsafe==2.1.5
    # via jinja2
marshmallow==3.22.0
    # via dataclasses-json
mccabe==0.7.0
    # via flake8
mock==5.1.0
    # via -r requirements.in
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
    #   shapely
orjson==3.10.7
    # via langsmith
packaging==24.1
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   langchain-core
    #   marshmallow
    #   pytest
pep8==1.7.1
    # via pytest-pep8
pluggy==1.5.0
    # via pytest
proto-plus==1.24.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-discoveryengine
    #   google-cloud-resource-manager
protobuf==5.28.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-discoveryengine
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pycodestyle==2.9.1
    # via flake8
pydantic==2.9.0
    # via
    #   -r requirements.in
    #   fastapi
    #   google-cloud-aiplatform
    #   langchain
    #   langchain-core
    #   langsmith
pydantic-core==2.23.2
    # via pydantic
pyflakes==2.5.0
    # via flake8
pyparsing==3.1.4
    # via httplib2
pytest==8.3.2
    # via
    #   -r requirements.in
    #   pytest-cache
    #   pytest-cov
    #   pytest-flake8
    #   pytest-pep8
pytest-cache==1.0
    # via pytest-pep8
pytest-cov==5.0.0
    # via -r requirements.in
pytest-flake8==1.2.2
    # via -r requirements.in
pytest-pep8==1.0.6
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via google-cloud-bigquery
python-dotenv==1.0.1
    # via uvicorn
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   responses
    #   uvicorn
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   langchain
    #   langchain-community
    #   langsmith
    #   responses
responses==0.25.3
    # via -r requirements.in
rsa==4.9
    # via google-auth
shapely==2.0.6
    # via google-cloud-aiplatform
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.34
    # via
    #   langchain
    #   langchain-community
starlette==0.38.5
    # via fastapi
tenacity==8.3.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   langchain-google-community
tomli==2.0.1
    # via
    #   coverage
    #   pytest
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   anyio
    #   fastapi
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   uvicorn
typing-inspect==0.9.0
    # via
    #   -r requirements.in
    #   dataclasses-json
tzdata==2024.1
    # via pydantic
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.2
    # via
    #   requests
    #   responses
uvicorn[standard]==0.30.6
    # via -r requirements.in
uvloop==0.20.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.0.1
    # via uvicorn
yarl==1.10.0
    # via aiohttp
